+++ app/Models/Message.php
@@ -20,11 +20,15 @@ class Message extends Model
 
     public function likes(): BelongsToMany
     {
-        return $this->belongsToMany(User::class, 'message_likes', 'message_uuid', 'user_id');
+        return $this->belongsToMany(User::class, 'message_likes', 'message_uuid', 'user_id', 'message_uuid', 'id');
     }
 
-        public function toggleLike($userId)
+    public function toggleLike($userId)
     {
+        if (!is_numeric($userId)) {
+            return response()->json(['error' => 'Invalid user ID'], 400);
+        }
+
         if ($this->likes()->where('user_id', $userId)->exists()) {
             $this->likes()->detach($userId);
             return 'Disliked';
@@ -33,7 +37,6 @@ class Message extends Model
             return 'Liked';
         }
     }
-
     public function incrementViews()
     {
         $this->increment('nb_vues');
